payments_system/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.1
djangorestframework==3.16.0
isort==6.0.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
